Moved IDirectoryProvider to the Quartz.Spi namespace, and moved the DefaultDirectoryProvider to the Quartz.Simpl namespace. Moved constructor to proper location in DirectoryScanJobModel. Updated the collection of updated files in DirectoryScanJob to be a thread-safe ConcurrentQueue. Updated the DirectoriesToScan collection on DirectoryScanJobModel to be a IReadOnlyList. Updated the DirectoryScanJobModel to run distinct on the collection returned by the IDirectoryProvider,so that each implementation does not need to remember to do that. Updated the DefaultDirectoryProvider to use both the "DIRECTORY_NAME" and "DIRECTORY_NAMES" merged JobDataMap values if both are provided, rather than only using "DIRECTORY_NAME". Added additional comments to the code. Updated some LINQ foreach method to normal foreach loops for clarity. Updated some logging code to use C#6 string interpolation.
